% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renorm.R
\name{anneal.tau}
\alias{anneal.tau}
\title{Simulated Annealing}
\usage{
anneal.tau(
  gg,
  e = NULL,
  v = NULL,
  vinv = NULL,
  L = NULL,
  t.lower,
  t.upper,
  start = NULL,
  complex = FALSE,
  method = c("eigen", "balanced", "square"),
  expm_method = c("Higham08.b"),
  tol = 1e-05,
  order = 1,
  print = FALSE,
  max_iter = 100,
  dt = 1,
  restarts = 25,
  cooling_rate = 0.99,
  seed = NULL
)
}
\arguments{
\item{gg}{original graph see \code{\link[=real.LRG]{real.LRG()}}}

\item{e}{Laplacian eigenvalues see \code{\link[=real.LRG]{real.LRG()}}}

\item{v}{Laplacian eigenvectors see \code{\link[=real.LRG]{real.LRG()}}}

\item{vinv}{Laplacian inverted eigenvectors see \code{\link[=real.LRG]{real.LRG()}}}

\item{L}{Laplacian see \code{\link[=real.LRG]{real.LRG()}}}

\item{t.lower}{minimal value for tau value search}

\item{t.upper}{maximal value for tau value search}

\item{complex}{should rho(tau) be treated as complex matrix see \code{\link[=real.LRG]{real.LRG()}}}

\item{method}{method of rho calculation see \code{\link[=real.LRG]{real.LRG()}}}

\item{expm_method}{method to compute exponential of the matrix
see \code{\link[=real.LRG]{real.LRG()}}}

\item{tol}{tolerance used to check if matrix for the exponential is
computationally singular see \code{\link[=real.LRG]{real.LRG()}}}

\item{order}{the Taylor approxymation order for the rho calculation see \code{\link[=real.LRG]{real.LRG()}}}

\item{print}{logical, does printing of intermediate results required}

\item{max_iter}{maximum of iterations}

\item{dt}{initial standard deviation for the search}

\item{restarts}{maximal number of restarts}

\item{cooling_rate}{cooling rate}

\item{seed}{random number generator seed}
}
\value{
list with the following slots:
\itemize{
\item df - \code{data.frame} with calculated statistics.
\item tau - vector of tested tau values.
\item plots - list of \code{\link[ggplot2]{ggplot}} objects for succesful
iterations.
\item plots2 - list of \code{\link[ggplot2]{ggplot}} objects for succesful
iterations.
\item restarts - number of restarts left.
}
}
\description{
Function calculates coarse-grained graph for selected set of tau values
and minimize the difference between degree distribution of the original
and coarse-grained graphs.
}
\seealso{
\code{\link[=real.LRG]{real.LRG()}}
}
