% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{annotate_go_mf}
\alias{annotate_go_mf}
\title{Add GO MF annotation to the graph vertices}
\usage{
annotate_go_mf(gg, annoF)
}
\arguments{
\item{gg}{graph to update}

\item{annoF}{annotation matrix in Pair form}
}
\value{
annotated igraph object
}
\description{
Function takes data from \code{annoF} matrix and add them to attributes
\code{GO_MF} for term and \code{GO_MF_ID} for IDs.
}
\examples{
file <- system.file("extdata", "PPI_Presynaptic.gml", package = "BioNAR")
gg <- igraph::read.graph(file,format="gml")
sfile<-system.file("extdata", "flatfile.go.MF.csv", package = "BioNAR")
goMF <- read.table(sfile,sep="\t",skip=1,header=FALSE,
strip.white=TRUE,quote="")
sgg <- annotate_go_mf(gg, goMF)
}
\seealso{
getAnnotationVertexList
}
