% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{annotate_SCHanno}
\alias{annotate_SCHanno}
\title{Add SCHanno synaptic functional groups}
\usage{
annotate_SCHanno(gg, anno)
}
\arguments{
\item{gg}{igraph object to annotate}

\item{anno}{annotation matrix in Pairs form}
}
\value{
annotated igraph object
}
\description{
Add SCHanno synaptic functional groups
}
\examples{
cid<-match('Presynaptic',getCompartments()$Name)
t<-getAllGenes4Compartment(cid)
gg<-buildFromSynaptomeByEntrez(t$HumanEntrez)
afile<-system.file("extdata", "SCH_flatfile.csv", package = "BioNAR")
dis    <- read.table(afile,sep="\t",skip=1,header=FALSE,
strip.white=TRUE,quote="")
agg<-annotate_SCHanno(gg,dis)
}
\seealso{
getAnnotationVertexList
}
