% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{annotate_vertex}
\alias{annotate_vertex}
\title{Generic annotation function}
\usage{
annotate_vertex(gg, name, values)
}
\arguments{
\item{gg}{igraph object to annotate}

\item{name}{name of the attribute}

\item{values}{annotation data.frame}
}
\value{
igraph object where vertex attribute \code{name} contains annotation
}
\description{
It takes name of the attribute, and two column
data.frame with vertex ID in the first column and annotation in the second.
As a first step all attributes with provided names will be removed.
}
\examples{
g1 <- make_star(10, mode="undirected")
V(g1)$name <- letters[1:10]
m<-data.frame(ID=letters[1:10],capital=LETTERS[1:10])
g2<-annotate_vertex(g1,vame='cap',values=m)
V(g2)$cap
}
