% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{annotateGeneNames}
\alias{annotateGeneNames}
\title{Annotate Human Gene Names}
\usage{
annotateGeneNames(gg)
}
\arguments{
\item{gg}{igraph object to annotate}
}
\value{
igraph object with new vertex attribute \code{GeneName}
}
\description{
For the protein-protein interaction (PPI) or disease gene interaction (DGN)
graphs that have EntrezID as a vertex name this function extract standard
name from \code{\link[org.Hs.eg.db]{org.Hs.eg.db}} and annotate vertices.
If for some
vertices EntresID does not match \code{\link[org.Hs.eg.db]{org.Hs.eg.db}}
empty string is added as GeneName.
}
\examples{
file <- system.file("extdata", "PPI_Presynaptic.gml", package = "BioNAR")
gg <- igraph::read.graph(file, format="gml")
agg<-annotateGeneNames(gg)
}
