% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{annotateGoBP}
\alias{annotateGoBP}
\title{Add GO BP annotation to the graph vertices}
\usage{
annotateGoBP(gg, annoF)
}
\arguments{
\item{gg}{graph to update}

\item{annoF}{annotation matrix in Pair form}
}
\value{
annotated igraph object
}
\description{
Function takes data from \code{annoF} matrix and add them to attributes
\code{GO_BP} for term and \code{GO_BP_ID} for IDs.
}
\examples{
file <- system.file("extdata", "PPI_Presynaptic.gml", package = "BioNAR")
gg <- igraph::read.graph(file, format="gml")
sfile<-system.file("extdata", "flatfile.go.BP.csv", package = "BioNAR")
goBP <- read.table(sfile, sep="\t", skip=1, header=FALSE,
strip.white=TRUE, quote="")
sgg <- annotateGoBP(gg, goBP)
}
\seealso{
getAnnotationVertexList
}
