% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{annotateGoCC}
\alias{annotateGoCC}
\title{Add GO CC  annotation to the graph vertices}
\usage{
annotateGoCC(gg, annoF)
}
\arguments{
\item{gg}{graph to update}

\item{annoF}{annotation matrix in Pair form}
}
\value{
annotated igraph object
}
\description{
Function takes data from \code{annoF} matrix and add them to attributes
\code{GO_CC} for term and \code{GO_CC_ID} for IDs.
}
\examples{
file <- system.file("extdata", "PPI_Presynaptic.gml", package = "BioNAR")
gg <- igraph::read.graph(file, format="gml")
sfile<-system.file("extdata", "flatfile.go.CC.csv", package = "BioNAR")
goCC <- read.table(sfile, sep="\t", skip=1, header=FALSE,
strip.white=TRUE, quote="")
sgg <- annotateGoCC(gg, goCC)
}
\seealso{
getAnnotationVertexList
}
