% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{annotateGoMF}
\alias{annotateGoMF}
\title{Add GO MF annotation to the graph vertices}
\usage{
annotateGoMF(gg, annoF, idatt = "name")
}
\arguments{
\item{gg}{graph to update}

\item{annoF}{annotation matrix in Pair form}
}
\value{
annotated igraph object
}
\description{
The function loads an annotation data matrix called \code{annoF}, which
contains three columns; the first containing gene Entrez IDs, the second gene
GO MF ID terms, the third gene GO MF description terms. The function then
performs a many-to-one mapping of each matrix row to a network vertex using
matching Entrez IDs, filling the vertices attributes \code{GO_MF_ID} and
\code{GO_MF}.
}
\examples{
file <- system.file("extdata", "PPI_Presynaptic.gml", package = "BioNAR")
gg <- igraph::read.graph(file, format="gml")
sfile<-system.file("extdata", "flatfile.go.MF.csv", package = "BioNAR")
goMF <- read.table(sfile, sep="\t", skip=1, header=FALSE,
strip.white=TRUE, quote="")
sgg <- annotateGoMF(gg, goMF)
}
\seealso{
getAnnotationVertexList
}
