% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{annotateInterpro}
\alias{annotateInterpro}
\title{Add InterPro Family and Domain annotation to the graph vertices}
\usage{
annotateInterpro(gg, annoF, annoD, idatt = "name")
}
\arguments{
\item{gg}{graph to update}

\item{annoF}{family annotation matrix in Pair form}

\item{annoD}{domain  annotation matrix in Pair form}

\item{idatt}{optional name of the vertex attributes that contains Entrez IDs}
}
\value{
annotated igraph object
}
\description{
Function takes data from \code{annoF} matrix and add them to attributes
\code{InterPro_Family} for term and \code{InterPro_Family_ID} for IDs.
}
\details{
Function takes data from \code{annoD} matrix and add them to attributes
\code{InterPro_Domain} for term and \code{InterPro_Domain_ID} for IDs.
}
\seealso{
getAnnotationVertexList
}
