% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{annotateSCHanno}
\alias{annotateSCHanno}
\title{Add SCHanno synaptic functional groups}
\usage{
annotateSCHanno(gg, anno)
}
\arguments{
\item{gg}{igraph object to annotate}

\item{anno}{annotation matrix in Pairs form}
}
\value{
annotated igraph object
}
\description{
Function adds Schizopherina related synaptic gene functional annotation
from Lips et al., (2012) \url{doi:10.1038/mp.2011.117} into \code{SCHanno}
vertex attribute.
}
\examples{
library(synaptome.db)
cid<-match('Presynaptic', getCompartments()$Name)
t<-getAllGenes4Compartment(cid)
gg<-buildFromSynaptomeByEntrez(t$HumanEntrez)
afile<-system.file("extdata", "SCH_flatfile.csv", package = "BioNAR")
dis    <- read.table(afile, sep="\t", skip=1, header=FALSE,
strip.white=TRUE, quote="")
agg<-annotateSCHanno(gg, dis)
}
\seealso{
getAnnotationVertexList
}
