% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{annotateTopOntoOVG}
\alias{annotateTopOntoOVG}
\title{Annotate graph with disease terms}
\usage{
annotateTopOntoOVG(gg, dis)
}
\arguments{
\item{gg}{igraph object to annotate}

\item{dis}{annotation matrix in Pairs form}
}
\value{
annotated igraph object
}
\description{
For the protein-protein interaction (PPI) or disease gene interaction (DGN)
graphs that have EntrezID as a vertex name this function takes annotation
from \code{dis} matrix and put it on the vertices with attributes
\code{TopOnto_OVG} for terms and \code{TopOnto_OVG_HDO_ID} for IDs.
}
\examples{
library(synaptome.db)
cid<-match('Presynaptic', getCompartments()$Name)
t<-getAllGenes4Compartment(cid)
gg<-buildFromSynaptomeByEntrez(t$HumanEntrez)
# read HDO data extracted from hxin/topOnto.HDO.db for synaptic network
afile<-system.file("extdata", "flatfile_human_gene2HDO.csv",
package = "BioNAR")
dis    <- read.table(afile, sep="\t", skip=1, header=FALSE,
strip.white=TRUE, quote="")
agg<-annotateTopOntoOVG(gg, dis)
}
\seealso{
getAnnotationVertexList
}
