% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{annotateVertex}
\alias{annotateVertex}
\title{Generic annotation function}
\usage{
annotateVertex(gg, name, values)
}
\arguments{
\item{gg}{igraph object to annotate}

\item{name}{name of the attribute}

\item{values}{annotation data.frame}
}
\value{
igraph object where vertex attribute \code{name} contains
annotation terms separated by semicolon.
}
\description{
It takes name of the attribute, and two column Pair form annotation
data.frame with vertex ID in the first column and annotation term in the
second. All terms annotating the same vertex ID will be collapsed with
semicolon as term separator.
}
\details{
As a first step all attributes with provided names will be removed.
}
\examples{
g1 <- make_star(10, mode="undirected")
V(g1)$name <- letters[1:10]
m<-rbind(data.frame(ID=letters[1:10], terms=letters[1:10]),
data.frame(ID=letters[1:10], terms=LETTERS[1:10]))
g2<-annotateVertex(g1, name='cap', values=m)
V(g2)$cap
}
\seealso{
getAnnotationVertexList
}
