% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BrProteins_2020.csv}
\alias{BrProteins_2020.csv}
\title{Title for BrProteins_2020.csv}
\description{
description for BrProteins_2020.csv
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{details for BrProteins_2020.csv
}\if{html}{\out{</div>}}
}
\keyword{BrProteins_2020.csv}
