% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNetwork.R
\name{buildFromSynaptomeByEntrez}
\alias{buildFromSynaptomeByEntrez}
\title{Utility function to create network from
\code{\link[synaptome.db]{synaptome.db}} data}
\usage{
buildFromSynaptomeByEntrez(entrez)
}
\arguments{
\item{entrez}{vector of EntrezIDs for network vertices}
}
\value{
largest connected component of network defined by the gene list
}
\description{
Utility function to create network from
\code{\link[synaptome.db]{synaptome.db}} data
}
\examples{
cid<-match('Presynaptic',getCompartments()$Name)
t<-getAllGenes4Compartment(cid)
gg<-buildFromSynaptomeByEntrez(t$HumanEntrez)
}
