% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNetwork.R
\name{buildFromSynaptomeGeneTable}
\alias{buildFromSynaptomeGeneTable}
\title{Utility function to create network from
\code{\link[synaptome.db]{synaptome.db}} data}
\usage{
buildFromSynaptomeGeneTable(t)
}
\arguments{
\item{t}{data.frame described in \code{\link[synaptome.db]{getGenesByID}}}
}
\value{
largest connected component of network defined by the gene table
}
\description{
Utility function to create network from
\code{\link[synaptome.db]{synaptome.db}} data
}
\examples{
cid<-match('Presynaptic',getCompartments()$Name)
t<-getAllGenes4Compartment(cid)
gg<-buildFromSynaptomeByEntrez(t)
}
