% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CorePSD95Complex.csv}
\alias{CorePSD95Complex.csv}
\title{Title for CorePSD95Complex.csv}
\description{
description for CorePSD95Complex.csv
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{details for CorePSD95Complex.csv
}\if{html}{\out{</div>}}
}
\keyword{CorePSD95Complex.csv}
