% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{diseasome.rda}
\alias{diseasome.rda}
\title{Title for diseasome.rda}
\description{
description for diseasome.rda
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{details for diseasome.rda
}\if{html}{\out{</div>}}
}
\keyword{diseasome.rda}
