% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{fitSigmoid}
\alias{fitSigmoid}
\title{Fit Fe distribution to sigmoid function}
\usage{
fitSigmoid(df, SDv)
}
\arguments{
\item{df}{enrichment \code{data.frame}}

\item{SDv}{vector of noise SD values}
}
\value{
list of fitted functions tables and plots
}
\description{
Fit Fe distribution to sigmoid function
}
