% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{flatfile.go.BP.csv}
\alias{flatfile.go.BP.csv}
\title{Title for flatfile.go.BP.csv}
\description{
description for flatfile.go.BP.csv
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{details for flatfile.go.BP.csv
}\if{html}{\out{</div>}}
}
\keyword{flatfile.go.BP.csv}
