% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{flatfile_human_gene2HDO.csv}
\alias{flatfile_human_gene2HDO.csv}
\title{Human Gene Disease Associations (GDA)}
\description{
Annotation derived from Human Disease Ontology database (HDO). The table
contains three columns; the first containing gene Entrez IDs, the second gene
Human Disease Ontology (HDO) ID terms, the third gene HDO description terms;
in csv format
}
\seealso{
\code{\link{annotateTopOntoOVG}}
}
\keyword{file}
