% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{getDiseases}
\alias{getDiseases}
\title{Get HDO disease IDs}
\usage{
getDiseases()
}
\value{
vector of disease IDs of interest
}
\description{
Return vector of HDO disease IDs for synaptic PPI analysis.
}
\examples{
getDiseases()
}
\seealso{
getDType
}
