% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{getDType}
\alias{getDType}
\title{Get DiseaseTypes}
\usage{
getDType()
}
\value{
vector of disease abbreviations for synaptic PPI analysis.
}
\description{
Return vector of disease abbreviations for synaptic PPI analysis.
}
\examples{
getDType()
}
\seealso{
getDiseases
}
