% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renorm.R
\name{getEigen}
\alias{getEigen}
\title{Calculate eigenvalues and eigenvectors of Laplacian matrix}
\usage{
getEigen(x, only.values = TRUE, inv.vec = FALSE)
}
\arguments{
\item{x}{Laplasian matrix from \code{\link{getLaplasian}}}

\item{only.values}{If TRUE only eigenvalues are returned.}

\item{inv.vec}{if TRUE inverted eigenvector matrix is returned.}
}
\value{
list with three slots: eigenvalues, eigenvectors and inverted
eigenvectors. If \code{only.values} is FALSE, both slots for eigenvectors and inverted
eigenvectors are NULL, if \code{inv.vec} is FALSE and \code{only.values} is TRUE, then
inverted eigenvectors slot is NULL.
}
\description{
Calculate eigenvalues and eigenvectors of Laplacian matrix
}
