% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{getIDs}
\alias{getIDs}
\title{Utility function to get vertex ids from vertex attributes
The function obtain attribute values and check duplicates in it.
It fails if any duplicate found.}
\usage{
getIDs(gg, idatt)
}
\arguments{
\item{gg}{graph}

\item{idatt}{attribute name}
}
\value{
\code{idatt} attribute values
}
\description{
Utility function to get vertex ids from vertex attributes
The function obtain attribute values and check duplicates in it.
It fails if any duplicate found.
}
