% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamo.R
\name{getPropagationMatrix}
\alias{getPropagationMatrix}
\title{Calculate propagate matrix from the topology and edge weights}
\usage{
getPropagationMatrix(g, attr = NULL)
}
\arguments{
\item{g}{igraph object}

\item{attr}{NULL or the name of edge attribute containing numerical weight
values}
}
\value{
sparce propagation matrix defined by the network topology and
edge attribute \code{attr} values
}
\description{
Algorithm proposed in:
}
\details{
Oron Vanunu, Oded Magger, Eytan Ruppin, Tomer Shlomi, Roded Sharan (2010)
"Associating Genes and Protein Complexes with Disease via Network Propagation"
PLoS Computational Biology \url{doi:10.1371/journal.pcbi.100064}
}
\examples{
data(karate, package='igraphdata')
w<-getPropagationMatrix(karate,attr='weight')
}
