% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{gofs}
\alias{gofs}
\title{Goodnes of fit KS test}
\usage{
gofs(x, rate, model, sigma2 = NULL, countDATA = TRUE)
}
\arguments{
\item{x}{steps along the Fe}

\item{rate}{parameters of the sigmoid}

\item{model}{fitted model}

\item{sigma2}{noise strength}

\item{countDATA}{should points to be counted}
}
\value{
list of \code{\link[stats]{ks.test}} values for each value in
\code{rate}
}
\description{
This is internal function and do not suppose to be called by user.
}
