% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPowerLaw.R
\docType{class}
\name{law-class}
\alias{law-class}
\title{Result of PawerLaw fit}
\description{
Result of PawerLaw fit
}
\section{Slots}{

\describe{
\item{\code{fit}}{\code{\link[poweRlaw]{displ-class}} result of power law fit.}

\item{\code{p}}{numeric.}

\item{\code{alpha}}{numeric degree of power-law.}

\item{\code{SDxmin}}{numeric bootstrap sd of Xmin.}

\item{\code{SDalpha}}{numeric bootstrap sd of alpha.}
}}

