% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasePairs.R
\name{permute}
\alias{permute}
\title{Randomly shuffle annotations}
\usage{
permute(GNS, N)
}
\arguments{
\item{GNS}{annotation list to take data from}

\item{N}{size of the sample}
}
\value{
random list of \code{GNS} values
}
\description{
This function is a convinience wrapper to \code{\link[base]{sample}} with
\code{replace= FALSE}
}
\examples{
permute(LETTERS,15)
}
