% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{plotRatio}
\alias{plotRatio}
\title{Plot fraction of enriched communities}
\usage{
plotRatio(
  xx,
  desc = "",
  anno = "",
  LEGtextSize = 1.5,
  LEGlineSize = 4,
  type = c("1", "2", "3", "4")
)
}
\arguments{
\item{xx}{enrichment statistics}

\item{desc}{plot subtitle}

\item{anno}{name of annotation used}

\item{LEGtextSize}{size of the text}

\item{LEGlineSize}{width of the line}

\item{type}{type of the plot}
}
\value{
ggplot object
}
\description{
Plot fraction of enriched communities
}
