% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{plotSigmoid}
\alias{plotSigmoid}
\title{Plot results of the sigmoid fit}
\usage{
plotSigmoid(x, rates, model, alg = "", pv = 0)
}
\arguments{
\item{pv}{}
}
\value{
\code{\link[ggplot2]{ggplot}} object with sigmoid fit plot
}
\description{
Plot results of the sigmoid fit
}
