% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PPI_cluster.gml}
\alias{PPI_cluster.gml}
\title{Title for PPI_cluster.gml}
\description{
description for PPI_cluster.gml
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{details for PPI_cluster.gml
}\if{html}{\out{</div>}}
}
\keyword{PPI_cluster.gml}
