% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PPI_Presynaptic.csv}
\alias{PPI_Presynaptic.csv}
\title{Title for PPI_Presynaptic.csv}
\description{
description for PPI_Presynaptic.csv
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{details for PPI_Presynaptic.csv
}\if{html}{\out{</div>}}
}
\keyword{PPI_Presynaptic.csv}
