% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PPI_Presynaptic.gml}
\alias{PPI_Presynaptic.gml}
\title{PPI graph for presynaptic compartment}
\description{
Protein-protein interactions (PPIS) for presynaptic compartment, extracted
from Synaptome.db, and saved in a graph format. Graph contains node
attributes, such as names (Entrez IDs), Gene Names, disease association
(TopOntoOVG, TopOntoOVGHDOID), annotation with schizophrenia-related genes
(Schanno (v/c), function annotation from GO (GOBPID, GOBP, GOMFID, GOMF,
GOCCID, GOCC), centrality measures (DEG - degree, BET - betweenness, CC -
clustering coefficient,  SL - semilocal centrality, mnSP - mean shortest
path, PR - page rank, sdSP - standard deviation of the shortest path), and
clustering memberships for 8 clustering algorithms (lec, wt, fc, infomap,
louvain, sgG1, sgG2, sgG5)
}
\keyword{file}
