% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{recluster}
\alias{recluster}
\title{Hierarchical graph clustering}
\usage{
recluster(GG, ALGN, CnMAX)
}
\arguments{
\item{GG}{graph to cluster}

\item{ALGN}{algorithm to apply}

\item{CnMAX}{maximum size of the cluster in \code{mem} that will not be
processed}
}
\value{
remembership matrix, that contains vertex ID membership and
result of reclustering
}
\description{
Function reads in a graph \code{GG} with cluster membership stored in vertex
attribute \code{ALGN}, and reapplies the clustering algorithm \code{ALGN} to
all clusters larger than \code{CnMAX}
}
\examples{
data(karate,package='igraphdata')
alg<-'louvain'
mem<-calcMembership(karate,alg = alg)
remem<-calcReclusterMatrix(karate,mem,alg,10)
}
