% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNetwork.R
\name{removeVertexTerm}
\alias{removeVertexTerm}
\title{Remove vertex property.}
\usage{
removeVertexTerm(GG, NAME)
}
\arguments{
\item{GG}{igraph object}

\item{NAME}{name of the vertex property to remove}
}
\value{
igraph object with attribute removed
}
\description{
Remove vertex property.
}
\examples{
data(karate,package='igraphdata')
vertex_attr_names(karate)
m<-removeVertexTerm(karate,'color')
vertex_attr_names(m)
}
