% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renorm.R
\name{runLRG}
\alias{runLRG}
\title{Facade function for the LRG analysis.}
\usage{
runLRG(
  gg,
  complex = FALSE,
  method = c("eigen", "balanced", "square"),
  expm_method = c("Higham08.b"),
  tol = 1e-05,
  order = 1,
  n.steps = 100,
  max_iter = 100,
  dt = 1,
  restarts = 25,
  cooling_rate = 0.85,
  seed = NULL,
  comb = "concat",
  simplify = TRUE,
  remove.loops = TRUE,
  remove.multiple = TRUE,
  verbatim = FALSE
)
}
\arguments{
\item{gg}{original graph}

\item{complex}{should rho(tau) be treated as complex matrix}

\item{method}{method of rho calculation}

\item{expm_method}{method to compute exponential of the matrix
\code{\link[expm]{expm}}}

\item{tol}{tolerance used to check if matrix for the exponential is
computationally singular \code{\link[expm]{expm}}}

\item{order}{the Taylor approxymation order for the rho calculation}

\item{comb}{Specifies how to combine the vertex attributes in the
coars-grained graph. Please see \code{\link[igraph]{contract}}
for details.}

\item{simplify}{logical, should coarse-grained graph be simplified at all.
Please see \code{\link[igraph]{simplify}} for details.}

\item{remove.loops}{logical, should self-loops be removed in the
coarse-grained graph.
Please see \code{\link[igraph]{simplify}} for details.}

\item{remove.multiple}{logical, should parallel edges be removed in the
coarse-grained graph.
Please see \code{\link[igraph]{simplify}} for details.}

\item{verbatim}{turns extensive logging on, do not use in Rmd or within other functions}
}
\description{
Facade function for the LRG analysis.
}
