% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasePairs.R
\name{runPermDisease}
\alias{runPermDisease}
\title{Calculate disease-disease pair overlaps on permuted network to estimate
its statistical significance}
\usage{
runPermDisease(
  gg,
  name,
  diseases = NULL,
  Nperm = 100,
  permute = c("random", "binned"),
  alpha = c(0.05, 0.01, 0.001)
)
}
\arguments{
\item{gg}{interactome network as igraph object}

\item{name}{name of the attribute that stores disease annotation}

\item{diseases}{list of diseases to match}

\item{Nperm}{number of permutations to apply}

\item{permute}{type of permutations.
\code{random} -- annotation is randomly shuffled,
\code{binned} -- annotation is shuffled in a way to preserve node
degree-annotation relationship by \code{\link{degree.binned.GDAs}}.}

\item{alpha}{statistical significance levels}
}
\value{
list of two matrices: \code{Disease_overlap_sig} gives s
tatistics for each pair of disease, and
\code{Disease_location_sig} gives intra-disease statistics
}
\description{
Function calculates disease overlap characteristics on the intact network
and then apply \code{Nperm} permutations of \code{permute} type. From
permuted networks function estimates significance of disease overlap and
store p-value, Bonferoni-adjusted p-value and q-value in the
\code{Disease_overlap_sig}.
Function also compares average disease separation of intact and
permuted network and calculate p-value with Wilcox test and store it in
\code{Disease_location_sig}.
}
\details{
Run with care, as large number of permutations could require a lot of
memory and be timeconsuming.
}
\examples{
file <- system.file("extdata", "PPI_Presynaptic.gml", package = "BioNAR")
gg <- igraph::read.graph(file,format="gml")
agg<-annotateGeneNames(gg)
r <- runPermDisease(
gg,
name = "TopOntoOVGHDOID",
diseases = c("DOID:10652","DOID:3312","DOID:12849","DOID:1826"),
Nperm = 10,
alpha = c(0.05, 0.01, 0.001))
r$Disease_location_sig
}
