% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasePairs.R
\name{sampleDegBinnedGDA}
\alias{sampleDegBinnedGDA}
\title{Perform degree-aware shuffling of the annotation.}
\usage{
sampleDegBinnedGDA(org.map, term)
}
\arguments{
\item{org.map}{degree-annotation mapping returned by
\code{\link{degreeBinnedGDAs}}}

\item{term}{annotation term to shuffle}
}
\value{
vertex IDs to assign \code{term} in shuffled annotation
}
\description{
Perform degree-aware shuffling of the annotation.
}
\examples{
file <- system.file("extdata", "PPI_Presynaptic.gml", package = "BioNAR")
gg <- igraph::read.graph(file, format="gml")
agg<-annotateGeneNames(gg)
gda<-prepareGDA(agg, 'TopOntoOVGHDOID')
diseases<-getAnnotationList(gda)
m<-degreeBinnedGDAs(agg, gda, diseases)
sampleDegBinnedGDA(m, diseases[1])
}
\seealso{
degreeBinnedGDAs
}
