% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renorm.R
\name{scan.tau}
\alias{scan.tau}
\title{Scan vector tau values for goodness of fit.}
\usage{
scan.tau(
  gg,
  e = NULL,
  v = NULL,
  vinv = NULL,
  L = NULL,
  tau,
  complex = FALSE,
  method = c("eigen", "balanced", "square"),
  expm_method = c("Higham08.b"),
  tol = 1e-05,
  order = 1,
  print = FALSE
)
}
\arguments{
\item{gg}{original graph see \code{\link[=real.LRG]{real.LRG()}}}

\item{e}{Laplacian eigenvalues see \code{\link[=real.LRG]{real.LRG()}}}

\item{v}{Laplacian eigenvectors see \code{\link[=real.LRG]{real.LRG()}}}

\item{vinv}{Laplacian inverted eigenvectors see \code{\link[=real.LRG]{real.LRG()}}}

\item{L}{Laplacian see \code{\link[=real.LRG]{real.LRG()}}}

\item{tau}{vector of tau values to calculate statistics on}

\item{complex}{should rho(tau) be treated as complex matrix see \code{\link[=real.LRG]{real.LRG()}}}

\item{method}{method of rho calculation see \code{\link[=real.LRG]{real.LRG()}}}

\item{expm_method}{method to compute exponential of the matrix
see \code{\link[=real.LRG]{real.LRG()}}}

\item{tol}{tolerance used to check if matrix for the exponential is
computationally singular see \code{\link[=real.LRG]{real.LRG()}}}

\item{order}{the Taylor approxymation order for the rho calculation see \code{\link[=real.LRG]{real.LRG()}}}

\item{print}{logical, does printing of intermediate results required}
}
\value{
list with the following slots:
\itemize{
\item df - \code{data.frame} with calculated statistics.
\item plots - list of \code{\link[ggplot2]{ggplot}} objects for succesful
iterations.
\item plots2 - list of \code{\link[ggplot2]{ggplot}} objects for succesful
iterations.
}
}
\description{
Function calculates coarse-grained graph for selected set of tau values
and compare degree distribution of the original and coarse-grained graphs.
}
\seealso{
\code{\link[=real.LRG]{real.LRG()}}
}
