% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SCH_flatfile.csv}
\alias{SCH_flatfile.csv}
\title{Title for SCH_flatfile.csv}
\description{
description for SCH_flatfile.csv
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{details for SCH_flatfile.csv
}\if{html}{\out{</div>}}
}
\keyword{SCH_flatfile.csv}
