% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{summaryStats}
\alias{summaryStats}
\title{Calculate summary statistics from enrichment table}
\usage{
summaryStats(RES, ALPHA, usePadj = FALSE, FeMAX = 0, FcMAX = 0)
}
\arguments{
\item{RES}{enrichment results \code{data.frame}}

\item{ALPHA}{p-value cut-off}

\item{usePadj}{logical, wether to use plain or adjusted p-value}

\item{FeMAX}{max of the FE}

\item{FcMAX}{max of the FC}
}
\value{
list of \code{data.frame}
}
\description{
Calculate summary statistics from enrichment table
}
