% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasePairs.R
\name{zeroNA}
\alias{zeroNA}
\title{Auxiliary function to replase NAs with zeros.}
\usage{
zeroNA(x)
}
\arguments{
\item{x}{matrix or vector to process}
}
\value{
matrix or vector with \code{NA}s replaced by zero.
}
\description{
Auxiliary function to replase NAs with zeros.
}
\examples{
x<-matrix(NA,nrow = 3,ncol = 3)
zeroNA(x)
}
