% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioNAR.R
\docType{package}
\name{BioNAR}
\alias{BioNAR}
\alias{BioNAR-package}
\title{BioNAR: Biological Network Analysis in R}
\description{
The R package BioNAR, developed to step by step analysis of PPI
network. The aim is to quantify and rank each protein’s simultaneous
impact into multiple complexes based on network topology and
clustering. Package also enables estimating of co-occurrence of
diseases across the network and specific clusters pointing towards
shared/common mechanisms.
}
\author{
\strong{Maintainer}: Anatoly Sorokin \email{lptolik@gmail.com}

Authors:
\itemize{
  \item Colin Mclean \email{Colin.D.Mclean@ed.ac.uk}
  \item Oksana Sorokina \email{oksana.sorokina@ed.ac.uk}
  \item J. Douglas Armstrong \email{Douglas.Armstrong@ed.ac.uk} [funder]
}

Other contributors:
\itemize{
  \item T. Ian Simpson \email{Ian.Simpson@ed.ac.uk} [contributor, funder]
}

}
