% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consMatrix.R
\name{buildConsensusMatrix}
\alias{buildConsensusMatrix}
\title{Build a consensus matrix from list of resampled clustering matrices outputted
from the function sampleGraphClust}
\usage{
buildConsensusMatrix(lcc)
}
\arguments{
\item{lcc}{list of clustering matrices obtained from the
\code{\link{sampleGraphClust}}}
}
\value{
consensus matrix
}
\description{
Build a consensus matrix from list of resampled clustering matrices outputted
from the function sampleGraphClust
}
