% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consMatrix.R
\name{buildConsensusMatrix}
\alias{buildConsensusMatrix}
\title{Build consensus matrix from the list of clusterings}
\usage{
buildConsensusMatrix(lcc)
}
\arguments{
\item{lcc}{list of clustering matrices obtained from the
\code{\link{sampleGraph}}}
}
\value{
consensus matrix
}
\description{
Build consensus matrix from the list of clusterings
}
