% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNetwork.R
\name{buildNetwork}
\alias{buildNetwork}
\title{Build network from data.table}
\usage{
buildNetwork(ff, kw = NA)
}
\arguments{
\item{ff}{network structure data.frame with first two columns defining the
network edge nodes}

\item{kw}{pmid keyword annotation data.frame. If \code{NA}
no annotation will be added}
}
\value{
igraph object of the largest connected component
}
\description{
It is a wrapper for \code{\link[igraph]{graph.data.frame}} that always
return allow annotate edges largest connected component of the network
defined by \code{ff}. It is also allows annotation of edges
with PubMed data, if exists.
}
\examples{
f<-data.frame(A=c('A','A','B','D'),B=c('B','C','C','E'))
gg<-buildNetwork(f)
V(gg)$name
}
