% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{calcAllClustering}
\alias{calcAllClustering}
\title{Calculate memberships for all clustering algorithms and store them on the
graph vertices.}
\usage{
calcAllClustering(gg)
}
\arguments{
\item{gg}{graph for analysis}
}
\value{
new graph object with all membership results stored as a vertex
attribute.
}
\description{
This function will call \code{\link{calcClustering}} for each clustering
algorithm given in our predefined list. In the event no clustering could be
performed, warnings will be issued and no new vertex attribute added to the
graph.
}
\examples{
g1 <- make_star(10, mode="undirected")
V(g1)$name <- letters[1:10]
g1<-calcAllClustering(g1)
clusteringSummary(g1)
}
\seealso{
calcClustering
}
