% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCentrality.R
\name{calcCentrality}
\alias{calcCentrality}
\title{Calculate the vertex centrality measures (degree, betweenness, closeness,
semi-local, etc....) for each graph vertex and store each result as
new vertex attribute in the graph.}
\usage{
calcCentrality(gg)
}
\arguments{
\item{gg}{igraph object}
}
\value{
modified igraph object
}
\description{
A wrapper function that first calls \code{\link{getCentralityMatrix}}, to
calculate all vertex centrality measures, and then
\code{\link{applpMatrixToGraph}} to store each centrality result as a new
vertex attribute in the graph.
}
\examples{
data(karate,package='igraphdata')
ggm<-calcCentrality(karate)
V(ggm)$DEG
}
