% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCentrality.R
\name{calcCentrality}
\alias{calcCentrality}
\title{Calculate centrality measures for graph nodes and save them as vertex
property.}
\usage{
calcCentrality(gg)
}
\arguments{
\item{gg}{igraph object}
}
\value{
modified igraph object
}
\description{
Wrapper finction that calls \code{\link{getCentralityMatrix}} to calculate
all available centrality
measires and \code{\link{applpMatrixToGraph}} to store them as a vertex
attributes.
}
\examples{
data(karate,package='igraphdata')
ggm<-calcCentrality(karate)
V(ggm)$DEG
}
