% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{calcClustering}
\alias{calcClustering}
\title{Calculate memberships for particular clustering algorithms and store
them on the graph vertices.}
\usage{
calcClustering(gg, alg)
}
\arguments{
\item{gg}{igraph object to cluster}

\item{alg}{algorithm to apply}
}
\value{
modified igraph object with calculated membership stored as a vertex
attribute and modularity as a graph attribute
}
\description{
Results of clustering algorithm application to the same graph could differ
between runs due to use of stochastic algorithm. To allow reproducible
downstream analysis clustering results are stored as graph attributes. This
function call \code{\link{getClustering}} and store membership as vertex
attribute and modularity as graph attribute with attribute name defined by
\code{alg} value.
}
\details{
NOTE: \code{\link{getClustering}} verifies algorithm names with
\code{\link[base]{match.arg}} so correct membership will be calculated, but
name of the attribute is taken from \code{alg} argument, so it is possible
that vertex attribute name won't exactly match name of the algorithm from
\code{link{getClustering}}.
}
\examples{
data(karate,package='igraphdata')
g<-calcClustering(karate,'louvain')
vertex_attr_names(g)
graph_attr(g,'louvain')
}
\seealso{
getClustering
}
