% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbationEntropy.R
\name{calcEntropy}
\alias{calcEntropy}
\title{Calculate the graph entropy for each perturbed vertex, and save the results
as new vertex attributes in the graph.}
\usage{
calcEntropy(gg, maxSr = NULL, exVal = NULL)
}
\arguments{
\item{gg}{igraph object}

\item{maxSr}{maxSr value, if NULL \code{getEntropyRate} will be called.}

\item{exVal}{expression values boundaries.
Two columns are expected: \code{xx} and \code{lambda}. If NULL
default values \code{c(2,14)} and \code{c(-14,14)} will be used
for \code{xx}  and \code{lambda} respectively.}
}
\value{
graph with SR_UP and SR_DOWN vertex attributes storing the graph
entropy values with over- or under-expressing each vertex.
}
\description{
Calculate the graph entropy for each perturbed vertex, and save the results
as new vertex attributes in the graph. Given a PPI network, the function will
calculate the graph entropy for each vertex over- and under-expressed, where
the user can set what over- and under-expressed values each vertex takes
(Teschendorff et al, 2014).
}
\examples{
library(synaptome.db)
cid<-match('Presynaptic', getCompartments()$Name)
t<-getAllGenes4Compartment(cid)
gg<-buildFromSynaptomeByEntrez(t$HumanEntrez)
gg<-annotateGeneNames(gg)
gg<- calcEntropy(gg)
}
