% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbationEntropy.R
\name{calcEntropy}
\alias{calcEntropy}
\title{Calculates perturbation entropy and save it as attribute on the graph.}
\usage{
calcEntropy(gg, maxSr = NULL, exVal = NULL)
}
\arguments{
\item{gg}{igraph object}

\item{maxSr}{maxSr value, if NULL \code{getEntropyRate} will be called.}

\item{exVal}{expression values boundaries.
Two columns are expected: \code{xx} and \code{lambda}. If NULL
default values \code{c(2,14)} and \code{c(-14,14)} will be used
for \code{xx}  and \code{lambda} respectively.}
}
\value{
graph with SR_UP and SR_DOWN attributes storing entropy values for
over- and underexpression respectively
}
\description{
Calculates perturbation entropy and save it as attribute on the graph.
}
\examples{
cid<-match('Presynaptic',getCompartments()$Name)
t<-getAllGenes4Compartment(cid)
gg<-buildFromSynaptomeByEntrez(t$HumanEntrez)
gg<-annotateGeneNames(gg)
gg<- calcEntropy(gg)
}
