% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{fitSigmoid}
\alias{fitSigmoid}
\title{Fit Fe distribution to sigmoid function}
\usage{
fitSigmoid(stat, SDv = c(0, 0.05, 0.1, 0.5))
}
\arguments{
\item{stat}{enrichment results}

\item{SDv}{vector of noise SD values}
}
\value{
list of fitted functions tables and plots
}
\description{
Grid plot is designed in a way to be viewed in the device at least 12 inches
in width and 12 inches in height.
}
