% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{flatfile.go.BP.csv}
\alias{flatfile.go.BP.csv}
\title{Annotation from Gene Ontology Biological Process (GO_BP)}
\description{
Annotation, downloaded from Gene Ontology for Biological Proceess domain.
The table has columns: the first containing gene gene functional group ID
terms, the second gene functional group description terms,
the third - Human gene Entrez IDs; in csv format
}
\seealso{
\code{\link{annotateGoBP}}
}
\keyword{file}
