% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{flatfile.go.CC.csv}
\alias{flatfile.go.CC.csv}
\title{Annotation from Gene Ontology Cellular Compartment (GO_CC)}
\description{
Annotation, downloaded from Gene Ontology for Cellular Compartment domain.
The table has columns: the first containing gene gene functional group ID
terms, the second gene functional group description terms,
the third - Human gene Entrez IDs; in csv format
}
\seealso{
\code{\link{annotateGoCC}}
}
\keyword{file}
