% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{flatfile.go.MF.csv}
\alias{flatfile.go.MF.csv}
\title{Annotation from Gene Ontology Molecular Function (GO_MF)}
\description{
Annotation, downloaded from Gene Ontology for Molecular Function domain.
The table has columns: the first containing gene gene functional group ID
terms, the second gene functional group description terms,
the third - Human gene Entrez IDs; in csv format
}
\seealso{
\code{\link{annotateGoMF}}
}
\keyword{file}
