% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCentrality.R
\name{getCentralityMatrix}
\alias{getCentralityMatrix}
\title{Calculate centrality measures for graph nodes.}
\usage{
getCentralityMatrix(gg)
}
\arguments{
\item{gg}{igraph object}
}
\value{
data.frame with following columns:
\itemize{
\item ID   - vertex ID
\item DEG  - degree
\item iDEG - in-degree (directed graph only)
\item oDEG - out-degree (directed graph only)
\item BET  - betweenness for undirected graph
\item dBET - betweenness when directionality is taken into account
(directed graph only)
\item CC   - clustering coefficient
\item SL   - semilocal centrality
\item mnSP - mean shortest path
\item PR   - page rank  for undirected graph
\item dPR  - page rank when directionality is taken into account
(directed graph only)
\item sdSP - standard deviation of the shortest path
}
}
\description{
Calculate centrality measures for graph nodes.
}
\examples{
library(synaptome.db)
cid<-match('Presynaptic',getCompartments()$Name)
t<-getAllGenes4Compartment(cid)
gg<-buildFromSynaptomeByEntrez(t$HumanEntrez)
m<-getCentralityMatrix(gg)
}
