% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCentrality.R
\name{getCentralityMatrix}
\alias{getCentralityMatrix}
\title{Calculate centrality measures for graph nodes.}
\usage{
getCentralityMatrix(gg)
}
\arguments{
\item{gg}{igraph object}
}
\value{
matrix with following columns:
\itemize{
\item ID - vertex ID
\item DEG - degree
\item BET - betweenness
\item CC - clustering coefficient
\item SL - semilocal centrality
\item mnSP - mean shortest path
\item PR - page rank
\item sdSP - standard deviation of the shortest path
}
}
\description{
Calculate centrality measures for graph nodes.
}
\examples{
library(synaptome.db)
cid<-match('Presynaptic',getCompartments()$Name)
t<-getAllGenes4Compartment(cid)
gg<-buildFromSynaptomeByEntrez(t$HumanEntrez)
m<-getCentralityMatrix(gg)
}
