% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbationEntropy.R
\name{getEntropy}
\alias{getEntropy}
\title{Calculates perturbation entropy}
\usage{
getEntropy(gg, maxSr = NULL, exVal = NULL)
}
\arguments{
\item{gg}{igraph object}

\item{maxSr}{maxSr value, if NULL \code{getEntropyRate} will be called.}

\item{exVal}{expression values boundaries.
Two columns are expected: \code{xx} and \code{lambda}.
If NULL default values \code{c(2,14)} and \code{c(-14,14)} will
be used for \code{xx} and \code{lambda} respectively.}
}
\value{
matrix with EntrexID, GeneNames, Degree, Enttropy UP and Down
}
\description{
Calculates perturbation entropy
}
\examples{
cid<-match('Presynaptic',getCompartments()$Name)
t<-getAllGenes4Compartment(cid)
gg<-buildFromSynaptomeByEntrez(t$HumanEntrez)
gg<-annotateGeneNames(gg)
e<- getEntropy(gg)
}
