% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbationEntropy.R
\name{getEntropyRate}
\alias{getEntropyRate}
\title{Calculate parameters for Entropy plot and calculations.}
\usage{
getEntropyRate(gg)
}
\arguments{
\item{gg}{igroph object}
}
\value{
list with values of maxSr and SRo
}
\description{
Calculate parameters for Entropy plot and calculations.
}
\examples{
data(karate,package='igraphdata')
ent <- getEntropyRate(karate)
}
