% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renorm.R
\name{getHmatrix}
\alias{getHmatrix}
\title{Calculates Magnetic Laplacian for directed network}
\usage{
getHmatrix(gg, g = 0.5)
}
\arguments{
\item{gg}{igraph object}

\item{g}{an electric charge parameter, which value should be in range
between 0 and 0.5.}
}
\value{
directed graph Magnetic Laplacian matrix
}
\description{
Algorithm is based upon following publication:
}
\details{
ref: https://arxiv.org/pdf/1606.08266
ref: https://doi.org/10.1016/j.acha.2017.01.004
Magnetic Eigenmaps for the Visualization of Directed Networks
Micha ̈el Fanuel, Carlos M. Ala ́ız, A ́ngela Fern ́andez, Johan A. K. Suykens
KU Leuven, Department of Electrical Engineering (ESAT), Kasteelpark Arenberg 10, B-3001 Leuven, Belgium
}
\examples{
data(foodwebs,package='igraphdata')
g<-foodwebs[['ChesLower']]
upgrade_graph(g)
H <- getHmatrix(g)
}
