% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renorm.R
\name{getLaplacian}
\alias{getLaplacian}
\title{Prepare Laplacian matrix from graph}
\usage{
getLaplacian(gg, el = 0.5, weights = NULL, type = c("igraph", "magnetic"))
}
\arguments{
\item{gg}{igraph object}

\item{el}{an electric charge parameter for \code{\link{getHmatrix}}}

\item{weights}{an optional weights for weighted Laplacian matrix}

\item{type}{type of Laplacian to calculate for directed graph}
}
\description{
Prepare Laplacian matrix from graph
}
\seealso{
getHmatrix
}
