% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memrob.R
\name{getRobustness}
\alias{getRobustness}
\title{Calculate cluster robustness from consensus matrix}
\usage{
getRobustness(gg, alg, conmat)
}
\arguments{
\item{gg}{igroph object}

\item{alg}{clustering algorithm}

\item{conmat}{consensus matrix}
}
\value{
data.frame that for each cluster \code{C} shows its size \code{Cn},
robustness \code{Crob} and robustness scaled to range \link{0,1}
\code{CrobScaled}.
}
\description{
Calculate cluster robustness from consensus matrix
}
\examples{
data(karate,package='igraphdata')
alg<-'louvain'
gg<-calcClustering(karate,alg = alg)
conmat<-makeConsensusMatrix(gg,N=100,mask = 10,alg = alg,type = 2)
clrob<-getRobustness(gg, alg = alg, conmat)
clrob
}
