% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{gofs}
\alias{gofs}
\title{Goodnes of fit KS test}
\usage{
gofs(x, rate, model, sigma2 = NULL, countDATA = TRUE)
}
\arguments{
\item{countDATA}{}
}
\value{
list of \code{\link[stats]{ks.test}} values for each value in
\code{rate}
}
\description{
Goodnes of fit KS test
}
