% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{layoutByCluster}
\alias{layoutByCluster}
\title{Calculate layout based upon membership}
\usage{
layoutByCluster(gg, mem, layout = layout_with_kk)
}
\arguments{
\item{gg}{graph to layout}

\item{mem}{membership data.frame from \code{\link{calcMembership}}}
}
\value{
Layout in a form of 2D matrx.
}
\description{
Function split graph into clusters and layout each cluster idependently.
}
\examples{
data(karate,package='igraphdata')
alg<-'louvain'
mem<-calcMembership(karate,alg = alg)
lay<-layoutByCluster(karate,mem)
#plot(karate,layout=lay)
}
