% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consMatrix.R
\name{makeConsensusMatrix}
\alias{makeConsensusMatrix}
\title{Function to build consensus matrix in memory}
\usage{
makeConsensusMatrix(
  gg,
  N = 500,
  mask = 20,
  alg,
  type,
  reclust = FALSE,
  Cnmax = 10
)
}
\arguments{
\item{gg}{graph to perturb}

\item{N}{number of perturbation steps}

\item{mask}{percentage of elements to perturbe}

\item{alg}{clustering alg.}

\item{type}{edges=>1 or nodes=>2  to mask}

\item{reclust}{logical to decide wether to invoke reclustering via
\code{\link{recluster}}}

\item{Cnmax}{maximus size of the cluster in \code{mem} that will not be
processed if reclustering is invoked}
}
\value{
consensus matrix of Nvert X Nvert
}
\description{
For assessing the robustness of the clustering randomization study could
be performed, which applies the same clustering algorithm to N perturbed
networks and returns the consensus matrix where each pair of nodes will be
assigned the probability to belong to the same cluster. Network is
perturbed by deleting \code{mask} percent of edges (\code{type=1}) or
nodes (\code{type=2}) from network before clustering.
}
\examples{
data(karate,package='igraphdata')
alg<-'louvain'
gg<-calcClustering(karate,alg = alg)
conmat<-makeConsensusMatrix(gg,N=100,mask = 10,alg = alg,type = 2)
dim(conmat)
}
