% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbationEntropy.R
\name{plotEntropy}
\alias{plotEntropy}
\title{Plot entropy values}
\usage{
plotEntropy(SRprime, subTIT = "Entropy", SRo = NULL, maxSr = NULL)
}
\arguments{
\item{SRprime}{results of \code{\link{getEntropy}} invocation}

\item{subTIT}{entropy axis label}

\item{SRo}{second element of results of
\code{\link{getEntropyRate}} invocation}

\item{maxSr}{first element of results of
\code{\link{getEntropyRate}} invocation}
}
\value{
ggplot2 object with diagram
}
\description{
Plot entropy values
}
\examples{
cid<-match('Presynaptic',getCompartments()$Name)
t<-getAllGenes4Compartment(cid)
gg<-buildFromSynaptomeByEntrez(t$HumanEntrez)
gg<-annotateGeneNames(gg)
ent <- getEntropyRate(gg)
SRprime <- getEntropy(gg, maxSr = NULL)
plotEntropy(SRprime, subTIT = "Entropy", SRo = ent$SRo, maxSr = ent$maxSr)
}
