% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{plotSigmoid}
\alias{plotSigmoid}
\title{Plot results of the sigmoid fit}
\usage{
plotSigmoid(x, rates, model, alg = "", pv = 0)
}
\arguments{
\item{x}{steps along the Fe}

\item{rates}{parameters of the sigmoid}

\item{model}{fitted model}

\item{alg}{name of the clustering algorithm}

\item{pv}{Kolmogorov-Smirnov test's p-value}
}
\value{
\code{\link[ggplot2]{ggplot}} object with sigmoid fit plot
}
\description{
Plot results of the sigmoid fit
}
