% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PPI_Presynaptic.csv}
\alias{PPI_Presynaptic.csv}
\title{Table of protein protein interactions for presynaptic compartment}
\description{
Protein-protein interactions (PPIS) for presynaptic compartment, extracted
from Synaptome.db, in a csv form. Columns A and B correspond to Entrez IDs
for interacting proteins A and B (node names); column We contains the edge
weights, if available.
}
\seealso{
\code{\link{buildNetwork}}
}
\keyword{file}
