% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasePairs.R
\name{prepareGDA}
\alias{prepareGDA}
\title{Get particular annotation from the graph in the Vertex Annotation form
and format it for further analysis.}
\usage{
prepareGDA(gg, name)
}
\arguments{
\item{gg}{igraph object to take annotation from}

\item{name}{name of the vertex attribute that contains annotation. If graph
vertices has no such attribute error is thrown.}
}
\value{
escaped annotation in Vertex Annotation form
}
\description{
Get particular annotation from the graph in the Vertex Annotation form
and format it for further analysis.
}
\examples{
file <- system.file("extdata", "PPI_Presynaptic.gml", package = "BioNAR")
gg <- igraph::read.graph(file,format="gml")
agg<-annotateGeneNames(gg)
gda<-prepareGDA(gg,'TopOntoOVGHDOID')
head(gda)
}
\seealso{
getAnnotationVertexList

escapeAnnotation
}
