% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PresynAn.csv}
\alias{PresynAn.csv}
\title{Presynaptic genes specific functional annotation}
\description{
Presynaptic genes functional annotation derived from Boyken at al. (2013)
\url{doi:10.1016/j.neuron.2013.02.027}. The table has columns: the first
containing  functional group ID terms, the second - gene functional group
description terms, third - gene Human Entrez Ids; in csv format
}
\seealso{
\code{\link{annotatePresynaptic}}
}
\keyword{file}
