% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renorm.R
\name{real.LRG}
\alias{real.LRG}
\title{Laplacian Renormalisation Group in real-space}
\usage{
real.LRG(
  e,
  v,
  vinv = NULL,
  L = NULL,
  t,
  gg,
  complex = FALSE,
  method = c("eigen", "balanced", "square"),
  expm_method = c("Higham08.b"),
  tol = 1e-05,
  order = 1
)
}
\arguments{
\item{e}{Laplacian eigenvalues from \code{\link{getEigen}}}

\item{v}{Laplacian eigenvectors from \code{\link{getEigen}}}

\item{vinv}{Laplacian inverted eigenvectors from \code{\link{getEigen}}}

\item{L}{Laplacian}

\item{t}{tau value for coarse-graining}

\item{gg}{original graph}

\item{complex}{should rho(tau) be treated as complex matrix}

\item{method}{method of rho calculation}

\item{expm_method}{method to compute exponential of the matrix \code{\link[expm]{expm}}}

\item{tol}{tolerance used to check if matrix for the exponential is computationally singular \code{\link[expm]{expm}}}

\item{order}{}
}
\description{
Laplacian Renormalisation Group in real-space
}
