% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampleGraphClust}
\alias{sampleGraphClust}
\title{Perturbe graph and calculate its clustering}
\usage{
sampleGraphClust(gg, mask = 20, alg, type, reclust = FALSE, Cnmax = 10)
}
\arguments{
\item{gg}{graph}

\item{mask}{percentage of elements to perturbe}

\item{alg}{clustering alg.}

\item{type}{edges=>1 or nodes=>2  to mask}

\item{reclust}{logical to decide wether to invoke reclustering via
\code{\link{recluster}}}

\item{Cnmax}{maximus size of the cluster in \code{mem} that will not be
processed if reclustering is invoked}
}
\value{
list of Nx3 matrices
}
\description{
Function delete \code{mask} percent of edges (\code{type=1}) or nodes
(\code{type=2}) from network find largest connected component and cluster it.
Result is stored as Nx3 matrix: first column is vertex ID, second column
is a flag shownig was vertex clustered (vertex ID) or not (-1), and a third
column is a membership value for clustered vertices or -1.
}
\details{
This is internal function and not supposed to be calle by end user.
}
\examples{
data(karate,package='igraphdata')
alg<-'louvain'
mem<-calcMembership(karate,alg = alg)
smpl<-BioNAR:::sampleGraphClust(karate,mask=10,alg,type=2)
}
