% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SCH_flatfile.csv}
\alias{SCH_flatfile.csv}
\title{Schizopherina related synaptic gene functional annotation.}
\description{
Annotation, manually curated from an external file: Lips et al., (2012)
doi:10.1038/mp.2011.117.The table has columns: the first
containing gene Human Entrez IDs, the second gene functional group ID terms,
the third gene functional group description terms; in csv format
}
\seealso{
\code{\link{annotateSCHanno}}
}
\keyword{file}
